Ajuste int por float

+++ Ordenamiento.py
@@ -59,8 +59,8 @@ def main():
             break
 
         else:
-            # Agregamos lo ingresado por el usuario, convertimos a entero el string para que funciones bien la comparacion en el metodo burbuja.
-            numeros.append(int(n))
+            # Agregamos lo ingresado por el usuario, convertimos a float el string para que funciones bien la comparacion en el metodo burbuja.
+            numeros.append(float(n))
 
     # Mostramos la lista de los numeros original.
     print(f'\nLa lista de numeros original es: {numeros}')
